Move FileLikeQueue to filielike_queue module
